feat : Added mini_printf() function and mini_exit_printf() function
Mini printf function, prints a string of character on the standard
output. It behaves the same way as the printf function i.e. it prints
when it encounter a \n or when the Buffer is full.

Notes : I still have to implement some error checking but as for now
it's okay