feat + fix : Added mini_fread function and fixed MYFILE struct to encapsulate it, added mini_fopen prototype.
As you can see, i've moved the typedef of MYFILE struct in the
mini_lib.h this achieve encapsulation as in glibc for the FILE type. It
essentially tells the end user that there is a MYFILE type, so he can
use it use function that return MYFILE type, but in the meantime the end
user can't tell how MYFILE behave or how it is implemented. I then hide
the implementation to the user but still allow him to use it through a
set of function, mini_fopen() mini_fread() etc...