Tests : Test of the functions mini_fopen(), mini_fread() and mini_fwrite.
At first I ran in some issue with the tests, turns out that when using
fwrite I was trying to write in file where I had not the right to. Atm
thoses tests make high use of glibc function but it's just for testing
right now, i'll replace it later.